20_08_20_before/Project/Final_Project/Basic/Basic-Practice2/requirement.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
get==2019.4.13
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
post==2019.4.13
pprint==0.1
public==2019.4.13
python-decouple==3.1
python-dotenv==0.10.3
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
urllib3==1.25.6
Werkzeug==0.16.0
